requirements.txt
face_recognition==1.3.0
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.13.0
pypdf2==3.0.1
pdfplumber==0.11.1
numpy==1.24.3
pillow==10.2.0
tensorflow-object-detection-api==0.1.1
cryptography==41.0.3
pandas==2.1.4
tabula-py==2.9.3
pytesseract==0.3.10
python-dotenv==1.0.0
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76

